refactor(cloudflare): split worker.js into src/calendar.js and src/index.js
No behavior change; entrypoint moved to src/index.js and public exports re-exported.
