+++ cloudflare/regybox-scheduler/wrangler.jsonc
@@ -1,7 +1,7 @@
 {
   "$schema": "node_modules/wrangler/config-schema.json",
   "name": "regybox-scheduler",
-  "main": "worker.js",
+  "main": "src/index.js",
   "compatibility_date": "2026-06-17",
   "preview_urls": false,
   "keep_vars": true,
